test(protocol): write comprehensive tests [wip]

+++ packages/protocol/contracts/Migrations.sol
@@ -1,9 +1,8 @@
 pragma solidity 0.5.10;
 
 contract Migrations {
-
     address public owner;
-    uint public lastCompletedMigration;
+    uint256 public lastCompletedMigration;
 
     modifier restricted() {
         if (msg.sender == owner) {
@@ -15,7 +14,7 @@ contract Migrations {
         owner = msg.sender;
     }
 
-    function setCompleted(uint _completed) public restricted {
+    function setCompleted(uint256 _completed) public restricted {
         lastCompletedMigration = _completed;
     }
